PAAC-319: Ensuring calculator is sent flexi-access (trigger date and post trigger amount) values.
