+++ app/views/pensionInputs.scala.html
@@ -1,4 +1,4 @@
-@(paacForm : Form[form.CalculatorFormFields], currentYear: String)(implicit request:Request[_])
+@(paacForm : Form[models.CalculatorFormFields], currentYear: String)(implicit request:Request[_])
 
 @import uk.gov.hmrc.play.views.html.helpers._
