+++ app/views/postTriggerPensionInputs.scala.html
@@ -1,4 +1,4 @@
-@(paacForm : Form[form.CalculatorFormFields], model: form.CalculatorFormFields)(implicit request:Request[_])
+@(paacForm : Form[models.CalculatorFormFields], model: models.CalculatorFormFields)(implicit request:Request[_])
 
 @import uk.gov.hmrc.play.views.html.helpers._
 @import models._
